pyproject.toml
[tool.poetry]
name = "news"
version = "0.1.0"
description = ""
authors = ["vaaghu <vignesh021102@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
beautifulsoup4 = "^4.12.3"
requests = "^2.31.0"
django = "^5.0.1"
django-extensions = "^3.2.3"
python-dateutil = "^2.8.2"
postgres = "^4.0"
pgvector = "^0.2.5"
djangorestframework = "^3.14.0"
pygments = "^2.17.2"
pandas = "^2.2.1"
pdfplumber = "^0.10.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
